calculated earnings

+++ Group1_F2019_MAD3004_FP/Employee/FixedBasedPartTime.swift
@@ -11,7 +11,7 @@ class FixedBasedPartTime:PartTime
 {
     var fixedAmount:Float
     
-    init(FixTEmpID: Int, FixTEmptype: String, FixTEmpName: String, FixTEmpAge: Int, FixTEmpEarning: Float, Rate: Float, HoursWorked: Int,FixedAmount:Float) {
+    init(FixTEmpID: Int, FixTEmptype: String, FixTEmpName: String, FixTEmpAge: Int, FixTEmpEarning: Float, Rate: Float, HoursWorked: Float,FixedAmount:Float) {
         self.fixedAmount=FixedAmount
         super.init(PTEmpID: FixTEmpID, PTEmptype: FixTEmptype, PTEmpName: FixTEmpName, PTEmpAge: FixTEmpAge, PTEmpEarning: FixTEmpEarning, Rate: Rate, HoursWorked: HoursWorked)
     }
